Fixed some folders
