+++ src/components/UserComponents/LoginComponents/LoginForm.tsx
@@ -1,0 +1,67 @@
+import React, { useState } from 'react';
+import { Button, ButtonGroup } from "@nextui-org/button";
+import { Card, CardHeader, CardBody, CardFooter } from "@nextui-org/card";
+import { Input } from "@nextui-org/input";
+import { useNavigate } from "react-router-dom";
+
+
+const LoginFormPopup: React.FC = () => {
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const navigate = useNavigate();
+
+  const handleLogin = () => {
+    // Handle login logic here
+    console.log('Logging in with', email, password);
+    navigate('/dashboard'); // Example navigation after login
+  };
+
+  return (
+    <div style={popupStyle}>
+      <Card style={{ width: '450px', background: colors.background }}>
+        <CardHeader>
+          <h3>Login</h3>
+        </CardHeader>
+        <CardBody>
+          <Input
+            fullWidth
+            isClearable
+            placeholder="Email"
+            value={email}
+            onChange={(e) => setEmail(e.target.value)}
+            style={{ marginBottom: '20px' }}
+          />
+          <Input
+            fullWidth
+            isClearable
+            type="password"
+            value={password}
+            onChange={(e) => setPassword(e.target.value)}
+            style={{ marginBottom: '20px' }}
+          />
+        </CardBody>
+        <CardFooter>
+          <ButtonGroup>
+            <Button onClick={handleLogin} style={{ background: colors.primary, color: colors.white }}>
+              Login
+            </Button>
+          </ButtonGroup>
+        </CardFooter>
+      </Card>
+    </div>
+  );
+};
+
+const popupStyle: React.CSSProperties = {
+  position: 'fixed',
+  top: '50%',
+  left: '50%',
+  transform: 'translate(-50%, -50%)',
+  background: colors.gradient,
+  padding: '20px',
+  boxShadow: `0 4px 8px ${colors.shadow.p30}`,
+  borderRadius: '8px',
+  zIndex: 1000,
+};
+
+export default LoginFormPopup;
