+++ src/components/UserComponents/RegisterComponents/RegisterForm.tsx
@@ -1,0 +1,6 @@
+//Primero importamos los componentes que necesitamos de la librería next.ui
+import {Button, ButtonGroup} from "@nextui-org/button";
+import {Card, CardHeader, CardBody, CardFooter} from "@nextui-org/card";
+import {Input} from "@nextui-org/input";
+import {useNavigate} from "react-router-dom";
+import { useState } from "react";
